+++ app/models/transfer/task_list.rb
@@ -42,7 +42,8 @@ class Transfer::TaskList < ::BaseTaskList
         identifier: :after_transfer,
         tasks: [
           Transfer::Task::ConfirmDateAcademyTransferredTaskForm,
-          Transfer::Task::RedactAndSendDocumentsTaskForm
+          Transfer::Task::RedactAndSendDocumentsTaskForm,
+          Transfer::Task::DeclarationOfExpenditureCertificateTaskForm
         ]
       }
     ]