+++ db/migrate/20240315085020_add_transfer_declaration_of_expenditure_certificate_task_attributes.rb
@@ -1,0 +1,8 @@
+class AddTransferDeclarationOfExpenditureCertificateTaskAttributes < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :declaration_of_expenditure_certificate_date_received, :date
+    add_column :transfer_tasks_data, :declaration_of_expenditure_certificate_correct, :boolean, default: false
+    add_column :transfer_tasks_data, :declaration_of_expenditure_certificate_saved, :boolean, default: false
+    add_column :transfer_tasks_data, :declaration_of_expenditure_certificate_not_applicable, :boolean, default: false
+  end
+end