+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add UI for "schools due to convert" export, which is replacing ESFA & Grant
   management and finance exports
+- Add the new "declaration of expenditure certificate task" to the transfers
+  task list
+- The "declaration of expenditure certificate date received" is included in the
+  "Grant management and finance unit" csv export
 
 ### Changed